+++ restaurant-ms/src/main/java/com/quick_bites/entity/RestaurantReview.java
@@ -17,7 +17,7 @@ public class RestaurantReview {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int reviewId;
+    private Long reviewId;
 
     @Range(min = 1, max = 5)
     private double rating;
