+++ restaurant-ms/src/main/java/com/quick_bites/services/review_service/GiveReviewToDish.java
@@ -4,6 +4,6 @@ import com.quick_bites.dto.review_dto.GiveReviewDto;
 
 public interface GiveReviewToDish {
 
-    void giveReview(GiveReviewDto reviewDto ,int dishId);
+    void giveReview(GiveReviewDto reviewDto ,Long dishId);
 
 }
